requirements.txt
fastapi==0.68.1
tortoise-orm==0.18.0
uvicorn